Ai_Hackathon/requirements.txt
ollama
response
pydantic_ai
json
langchain_community
langchain_openai
langchain_core
langchain_pinecone
python_dotenv
pinecone_client
unstructured
markdown
langchain_nomic
langchain_elasticsearch
langchain_ollama
beautifulsoup4
tiktoken
faiss-cpu
langchain-ollama